Remove generic test framework
The framework ended up being too inflexible for new kinds of annotations. In
particular, dealing with LD was problematic.

It also did not have good granularity of tasks and was not flexible enough to
adjust.

Finally, the general method for intersecting, etc. was further developed
separately for T1D and diverged greatly. We found generalizing from that code
was easier and had nicer properties (granularity, simplicity).