Generate negative control for cutoffs enrichment
Generate random cell types from existing annotations by: 1) sampling the number
of elements from the distribution of number of elements across existing cell
types; 2) sampling that many elements from all the elements in the existing
annotations.

This tests whether the identity of which elements are in which cell types is
the relevant one; we expect that by destroying this association the result will
disappear.